ばしょのコンテンツ削除時に確認ダイアログを表示するように変更

+++ src/app/components/areas/areas.component.ts
@@ -4,6 +4,7 @@ import { MatDialog } from '@angular/material/dialog';
 
 import { AreasService } from '../../services/areas.service';
 import { AreasDialog } from '../dialogs/areas.dialog';
+import { ConfirmDialog } from '../dialogs/confirm.dialog';
 
 @Component({
   selector: 'app-areas',
@@ -26,8 +27,14 @@ export class AreasComponent implements OnInit {
   clickEditButton(key:string, name:string){
     this.openDialog(false, key, name);
   }
-  clickDeleteButton(key:string): void {
-    this.service.delete(key);
+  clickDeleteButton(key:string, name:string): void {
+    const confirmDialogRef = this.dialog
+    .open(ConfirmDialog, {
+      data: {name:name, action:'さくじょ'},
+    });
+    confirmDialogRef.afterClosed().subscribe(res => {
+      if(res) this.service.delete(key);
+    });
   }
 
   private openDialog(isNew:boolean, key:string|null, name:string){
